+++ packages/api/package.json
@@ -10,27 +10,37 @@
   "types": "./dist/index.d.ts",
   "license": "MIT",
   "dependencies": {
+    "@envelop/apollo-server-errors": "1.0.2",
+    "@envelop/core": "1.0.3",
+    "@envelop/depth-limit": "1.0.0",
+    "@envelop/disable-introspection": "1.0.0",
+    "@envelop/filter-operation-type": "1.0.0",
+    "@envelop/parser-cache": "1.0.0",
+    "@envelop/validation-cache": "1.0.0",
     "@graphql-tools/merge": "7.0.0",
+    "@graphql-tools/schema": "8.0.3",
+    "@graphql-tools/utils": "8.0.2",
     "@prisma/client": "2.29.1",
     "@types/pino": "6.3.11",
     "apollo-server-lambda": "2.25.2",
     "core-js": "3.16.1",
     "crypto-js": "4.1.1",
     "graphql": "15.5.1",
-    "graphql-depth-limit": "1.1.0",
+    "graphql-helix": "1.7.0",
+    "graphql-playground-html": "1.6.29",
     "graphql-scalars": "1.10.0",
     "jsonwebtoken": "8.5.1",
     "jwks-rsa": "2.0.4",
     "lodash.merge": "4.6.2",
     "lodash.omitby": "4.6.0",
-    "pino": "6.13.0",
     "pino-pretty": "5.1.3",
+    "pino": "6.13.0",
     "uuid": "8.3.2"
   },
   "devDependencies": {
+    "@envelop/types": "1.0.2",
     "@redwoodjs/auth": "0.36.0",
     "@types/crypto-js": "4.0.2",
-    "@types/graphql-depth-limit": "1.1.2",
     "@types/jsonwebtoken": "8.5.4",
     "@types/lodash.merge": "4.6.6",
     "@types/lodash.omitby": "4.6.6",
