+++ packages/api/src/errors.ts
