--- packages/graphql-server/__mocks__/@prisma/client.js
@@ -1,1 +1,0 @@
-export const PrismaClient = class MockPrismaClient {}
