+++ packages/graphql-server/jest.config.js
@@ -1,5 +1,1 @@
 module.exports = {}
-
-process.env = Object.assign(process.env, {
-  WEBHOOK_SECRET: 'MY_VOICE_IS_MY_PASSPORT_VERIFY_ME',
-})
