+++ packages/graphql-server/src/__tests__/graphql-server.test.ts
@@ -1,0 +1,5 @@
+describe('graphql-server', () => {
+  it('is deprecated', () => {
+    expect(true).toBeTruthy()
+  })
+})
