+++ tasks/e2e/cypress/integration/01-tutorial/codemods/Step0_2_GraphQL.js
@@ -11,7 +11,7 @@ import { logger } from 'src/lib/logger'
 import services from 'src/services/**/*.{js,ts}'
 
 export const handler = createGraphQLHandler({
-  loggerConfig: { logger, options: { data: true, query: true } },
+  loggerConfig: { logger, options: { tracing: true, userAgent: true, data: true, query: true } },
   schema: makeMergedSchema({
     schemas,
     services: makeServices({ services }),
